chore: separate testimonials component
