+++ src/app/ai/page.jsx
@@ -2,10 +2,10 @@ import Examples from 'components/pages/ai/examples';
 import Hero from 'components/pages/ai/hero';
 import Integration from 'components/pages/ai/integration';
 import Stats from 'components/pages/ai/stats';
-// import Testimonials from 'components/pages/ai/testimonials';
 import CTAWithElephant from 'components/shared/cta-with-elephant';
 import Layout from 'components/shared/layout';
 import SplitViewGrid from 'components/shared/split-view-grid';
+// import Testimonials from 'components/shared/testimonials';
 import LINKS from 'constants/links';
 import SEO_DATA from 'constants/seo-data';
 import compatibilityIcon from 'icons/ai/compatibility.svg';
@@ -39,6 +39,20 @@ const items = [
   },
 ];
 
+// TODO: add links to original tweets?
+// const sliderItems = [
+//   {
+//     text: 'What if PGVector was on steroids ?! <a href="https://twitter.com/raoufdevrel" target="_blank" rel="noreferrer noopener">@raoufdevrel</a> from <a href="https://twitter.com/neondatabase" target="_blank" rel="noreferrer noopener">@neondatabase</a> just did that with pg_embedding and made it available through <a href="https://twitter.com/LangChainAI" target="_blank" rel="noreferrer noopener">@LangChainAI</a>. I&apos;ll definitely give it a try on <a href="https://twitter.com/quivr_brain" target="_blank" rel="noreferrer noopener">@quivr_brain</a> and our 10 GiB of vectors that use PGVector',
+//     authorName: 'Stan Girard',
+//     authorTitle: 'Founder, Quivr',
+//   },
+//   {
+//     text: '<a href="https://twitter.com/postgresql" target="_blank" rel="noreferrer noopener">@PostgreSQL</a> is popular database choice. Excited to share a new extension from <a href="https://twitter.com/neondatabase" target="_blank" rel="noreferrer noopener">@neondatabase</a> to help you use it for embeddings as well (with HNSW)!',
+//     authorName: 'Harrison Chase',
+//     authorTitle: 'Co-Founder and CEO, LangChainAI',
+//   },
+// ];
+
 const AIPage = () => (
   <Layout
     className="bg-black-new text-white"
@@ -65,7 +79,12 @@ const AIPage = () => (
     />
     <Examples />
     {/* removed testimonials section for now as part of removing the pg_embedding references */}
-    {/* <Testimonials /> */}
+    {/* <Testimonials
+      className="mt-40 xl:mt-[120px] lg:mt-28 md:mt-20"
+      itemClassName="min-h-[222px] md:min-h-[184px] sm:min-h-[252px]"
+      items={sliderItems}
+      theme="twitter"
+    /> */}
     <CTAWithElephant
       className="mt-[180px] 2xl:mt-40 xl:mt-[125px] lg:mt-16"
       buttonClassName="px-14 xl:px-10 lg:px-9 sm:px-14"
