+++ src/components/shared/testimonials/slider-item/index.js
